// src/renderer/features/quan-ly-lop-hoc-phan/types.ts
export interface QuanLyLopHocPhanFilters {
  idBacDaoTao?: string;
  idKhoaChuQuan?: string;
  idCoSo?: string;
  idKhoaHoc?: string;
  idLoaiDaoTao?: string;
  idNganhHoc?: string;
  idChuyenNganh?: string;
  idLopDanhNghia?: string;
  lopDuKien?: string;
  monHocLopHocPhan?: string;
  gvDuKien?: string;
  idTrangThai?: string;
  loaiLop?: string;
  idHinhThucThi?: string;
  idKhoaChuQuanLop?: string;
}